fix: preserve compact context gate escapes
